+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-- Fixed it could show board label in wrong position in schematric image
+- Fixed it showed some board labels in wrong position in schematric image
 - Fixed it did not save main region selection
 - Fixed `ESCAPE` did not close active component info popup, if set to "Open multiple windows for popup" #84
 - Added possibility to reorder schematic thumbnails with mouse drag/drop